Move UKF sigma points to sampling
