+++ docs/api-overview.md
@@ -71,6 +71,7 @@ Samplers and grid generators for Euclidean and manifold domains.
 Common starting points include:
 
 - `GaussianSampler`;
+- `MerweScaledSigmaPoints` and `JulierSigmaPoints`;
 - `FibonacciGridSampler`;
 - `SphericalFibonacciSampler`;
 - `HealpixHopfSampler`;
