+++ src/pyrecest/filters/unscented_kalman_filter.py
@@ -4,8 +4,8 @@ from typing import Callable
 import pyrecest.backend
 from pyrecest.backend import atleast_1d, zeros
 from pyrecest.distributions import GaussianDistribution
+from pyrecest.sampling.sigma_points import MerweScaledSigmaPoints
 
-from ._ukf import MerweScaledSigmaPoints
 from ._ukf import UnscentedKalmanFilter as BayesianFiltersUKF
 from ._ukf import _UKFModel
 from .abstract_filter import AbstractFilter
